backend/pyproject.toml
[tool.isort]
profile = "black"

[tool.poetry]
name = "swoops"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "~3.9"
Django = "4.0.5"
django-configurations = "2.3.2"
dj-database-url = "0.5.0"
psycopg2-binary = "2.9.6"
python-dotenv = "0.20.0"
whitenoise = {extras = ["brotli"], version = "6.2.0"}
gunicorn = "20.1.0"
djangorestframework = "3.13.1"
django-cors-headers = "3.13.0"
django-filter = "21.1"
sentry-sdk = "1.6.0"
requests = "2.28.0"
py-ecc = "6.0.0"
eth-hash = {extras = ["pycryptodome"], version = "0.3.3"}
django-pglocks = "1.0.4"
web3 = "6.0.0b4"
django-pgtrigger = "2.4.1"
django-pghistory = "1.5.0"
redis = "4.3.4"
celery = "5.2.7"
django-dynamic-fixture = "3.1.2"
django-pgbulk = "1.2.0"
djangorestframework-simplejwt = "5.2.0"
django-otp = "^1.1.3"
qrcode = "^7.3.1"
django-axes = "^5.39.0"
django-admin-colors = "^0.2.0"
django-extensions = "^3.2.1"
boto3 = "^1.25.4"
moto = {extras = ["s3"], version = "^4.0.8"}
pillow = "^9.3.0"
sendgrid = "^6.9.7"
django-debug-toolbar = "^3.8.1"
numpy = "^1.24.2"
opencv-python-headless = "^4.7.0.72"
mypy = "^1.1.1"
scout-apm = "2.26.1"
friendlywords = "^1.1.2"
etherscan-python = "^2.1.0"

[tool.poetry.dev-dependencies]
pre-commit = "2.20.0"
drf-yasg = "1.21.3"
pytest-django = "4.5.2"
pytest-dotenv = "0.5.2"
django-dynamic-fixture = "3.1.2"
black = "22.12.0"
isort = "5.10.1"
pytest = "7.1.2"
flake8 = "4.0.1"
pytest-mock = "3.8.2"
freezegun = "1.2.1"
moto = "^4.0.8"
requests-mock = "^1.10.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''
